+++ src/image_manager/Puller.hpp
@@ -41,7 +41,7 @@ private:
     std::string getServerUri(const std::string &server);
     void saveImage(web::json::value fsLayers);
     void saveLayer(const std::string &digest);
-    void downloadStream(const std::string &uri, const std::string &path, const boost::filesystem::path &filename);
+    void downloadStream(const web::http::http_response &response, const boost::filesystem::path &filename);
     std::string requestAuthorizationToken(web::http::http_response& response);
     bool checkSum(const std::string &digest, const boost::filesystem::path &filename);
     void printLog(  const boost::format &message, common::LogLevel LogLevel,